BackEnd/requirements.txt
amqp==5.1.0
aniso8601==9.0.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
Flask-RESTful==0.3.9
flower==1.0.0
humanize==4.0.0
itsdangerous==2.1.1
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2022.1
redis==4.1.4
six==1.16.0
sqlvalidator==0.0.18
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
pyjwt==2.0.0
